# Speed up grouping of indentifiers