This should help with https://github.com/andialbrecht/sqlparse/issues/135 and https://github.com/andialbrecht/sqlparse/issues/199. Using the file uploaded to https://github.com/andialbrecht/sqlparse/issues/199, this branch reformats the sql 10x faster:
```
# master
time sqlformat -r ./test.sql -o /dev/null
sqlformat -r ./test.sql -o /dev/null  9.61s user 0.03s system 99% cpu 9.663 total

# this PR
time sqlformat -r ./test.sql -o /dev/null
sqlformat -r ./test.sql -o /dev/null  0.95s user 0.02s system 99% cpu 0.979 total
```